% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxivistr.R
\name{article_details}
\alias{article_details}
\title{Article details}
\usage{
article_details(ID)
}
\arguments{
\item{ID}{Rxivist paper ID.}
}
\value{
Returns a list with details about the specified article.
}
\description{
Retrieves data about a single paper and all of its authors
}
\examples{
article_details(72514)
}
