% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxivistr.R
\name{article_downloads}
\alias{article_downloads}
\title{Article downloads}
\usage{
article_downloads(ID)
}
\arguments{
\item{ID}{Rxivist paper ID.}
}
\value{
Returns a data frame with information about monthly downloads for the specified article.
}
\description{
Returns monthly download statistics for articles.
}
\examples{
article_downloads(72514)
}
