% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxivistr.R
\name{author}
\alias{author}
\title{author}
\usage{
author(ID)
}
\arguments{
\item{ID}{Rxivist author ID.}
}
\value{
Returns a data frame with information about the specified author.
}
\description{
Provides information about the specified author.
}
\examples{
author(206102)
}
