% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxivistr.R
\name{authors_rank}
\alias{authors_rank}
\title{authors_rank}
\usage{
authors_rank(category = "")
}
\arguments{
\item{category}{Filters out results not related to the specified category. Category list can be viewed with category_list function. If filtering for a single category, this category can be specified as a string. If filtering for multiple categories, the argument must be specified as a vector, e.g. category = c("zoology","biophysics"). If it is left as default, the function will return top 200 authors overall.}
}
\value{
Returns a data frame with the top 200 authors.
}
\description{
Returns top 200 authors in specified category.
}
\examples{
authors_rank("plant-biology")
}
