% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxivistr.R
\name{category_list}
\alias{category_list}
\title{Category_list}
\usage{
category_list()
}
\value{
Returns a list with all article categories.
}
\description{
Lists all categories
}
\examples{
category_list()
}
