% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxivistr.R
\name{check_args}
\alias{check_args}
\title{check_args}
\usage{
check_args(from, category, sortby)
}
\arguments{
\item{from}{from argument}

\item{category}{category argument}

\item{sortby}{sortby argument}
}
\value{
Stops main function.
}
\description{
Check arguments.
}
