% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxivistr.R
\name{rxivist_stats}
\alias{rxivist_stats}
\title{rxivist_stats}
\usage{
rxivist_stats()
}
\value{
Returns a list with rxivist stats.
}
\description{
Returns statistics on number of articles indexed by rxivist.
}
\examples{
rxivist_stats()
}
